# packages/backend/api-docs.yaml
openapi: 3.1.0
info:
  title: Tienda Sol API
  version: 1.0.0
  description: API REST de Tienda Sol.
servers:
  - url: http://localhost:3000/api/v1
...
security:
  - userIdAuth: []
paths:
...
  /pedidos/{id}:
    get:
      tags: [pedidos]
      operationId: consultarPedido
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pedido' }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404': { $ref: '#/components/responses/Error404' }
        '422': { $ref: '#/components/responses/Error422' }
    patch:
      tags: [pedidos]
      operationId: cambiarEstadoPedido
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CambioEstado' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pedido' }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404': { $ref: '#/components/responses/Error404' }
        '409': { $ref: '#/components/responses/Error409' }
        '422': { $ref: '#/components/responses/Error422' }
...
components:
  securitySchemes:
    userIdAuth:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    Error400:
      description: Error de validación
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Error401:
      description: No autenticado
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Error403:
      description: Prohibido
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Error404:
      description: No encontrado
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Error409:
      description: Conflicto
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Error422:
      description: Error de dominio
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    Moneda:
      type: string
      enum: [PESO_ARG, DOLAR_USA, REAL]
...
    EstadoPedido:
      type: string
      enum: [cancelado, entregado, enviado, enPreparacion, confirmado, pendiente]
...
    ItemPedido:
      type: object
      required: [productoID, cantidad, precioUnitario, subtotal]
      properties:
        productoID: { type: integer }
        cantidad: { type: integer }
        precioUnitario: { type: number }
        subtotal: { type: number }
    DireccionEntrega:
      type: object
      required: [calle, altura, codigoPostal, ciudad, provincia, pais]
      properties:
        calle: { type: string }
        altura: { type: integer }
        piso: { type: string, nullable: true }
        departamento: { type: string, nullable: true }
        codigoPostal: { type: string }
        ciudad: { type: string }
        provincia: { type: string }
        pais: { type: string }
        lat: { type: number, nullable: true }
        lon: { type: number, nullable: true }
...
    CambioEstado:
      type: object
      required: [nuevoEstado, motivo]
      properties:
        nuevoEstado: { $ref: '#/components/schemas/EstadoPedido' }
        motivo: { type: string, minLength: 1 }
    Pedido:
      type: object
      required: [id, compradorId, vendedorId, estado, items, total, moneda, direccionEntrega]
      properties:
        id: { type: integer }
        compradorId: { type: integer }
        vendedorId: { type: integer }
        estado: { $ref: '#/components/schemas/EstadoPedido' }
        items:
          type: array
          items: { $ref: '#/components/schemas/ItemPedido' }
        moneda: { $ref: '#/components/schemas/Moneda' }
        direccionEntrega: { $ref: '#/components/schemas/DireccionEntrega' }
        historialEstados:
          type: array
          items:
            type: object
            properties:
              estado: { $ref: '#/components/schemas/EstadoPedido' }
              fecha: { type: string, format: date-time }
              motivo: { type: string, nullable: true }
        total: { type: number }
        fechaCreacion: { type: string, format: date-time, nullable: true }
        fechaActualizacion: { type: string, format: date-time, nullable: true }
        fechaCancelacion: { type: string, format: date-time, nullable: true }
        fechaEnvio: { type: string, format: date-time, nullable: true }
        fechaEntrega: { type: string, format: date-time, nullable: true }
...
    Error:
      type: object
      properties:
        error: { type: string }
        detalles: { nullable: true }